iot_project/requirements.txt
django
djangorestframework
sqlalchemy
pytest
pytest-django
pytest-mock
mysqlclient
pymysql
django-cors-headers